requirements.txt
streamlit
langchain
langchain_google_genai
pypdf2
faiss-cpu
python-dotenv
pandas